pyproject.toml
[tool.poetry]
name = "minecat"
version = "0.3.0"
description = "A Discord bot to connect with Minecraft!"
authors = ["ooliver1 <oliverwilkes2006@icloud.com>"]
license = "ELv2"

[tool.poetry.dependencies]
python = "^3.10"
nextcord = {extras = ["speed"], version = "^2.0.0"}
aiohttp = "^3.8.1"
websockets = "^10.3"
uvloop = "^0.16.0"
orjson = "^3.7.8"
lru-dict = "^1.1.7"
ooliver-botbase = "^1.20.0"
cachetools = "^5.2.0"
nextcord-ext-ipc = "^2.2.1"
nextcord-ext-menus = "^1.5.4"
asyncpg = "^0.26.0"
shortuuid = "^1.0.9"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
